fix for faker name added dynatable added few styles fixed some rspec errors
!!! admin can also destroy/edit - its not a bug, its a feature!
    admin should have possibility of doing such things, its not a category-maker
    -> rspec can fail because of that

fixed review-user relationship (it didnt save user_id in database
changed classes for displaying notices/errors

I tried to make it look better, but im not a designer ;_;
zzz... Noone will read it.

+++ app/controllers/categories_controller.rb
@@ -12,32 +12,50 @@ class CategoriesController < ApplicationController
   end
 
   def new
+    if !current_user.admin?
+      redirect_to new_user_session_path
+    end
   end
 
   def edit
+    if !current_user.admin?
+      redirect_to new_user_session_path
+    end
   end
 
   def create
-    self.category = Category.new(category_params)
-
-    if category.save
-      redirect_to category, notice: 'Category was successfully created.'
+    if !current_user.admin?
+      redirect_to new_user_session_path
     else
-      render action: 'new'
+      self.category = Category.new(category_params)
+  
+      if category.save
+        redirect_to category, notice: 'Category was successfully created.'
+      else
+        render action: 'new'
+      end      
     end
   end
 
   def update
-    if category.update(category_params)
-      redirect_to category, notice: 'Category was successfully updated.'
+    if !current_user.admin?
+      redirect_to new_user_session_path
     else
-      render action: 'edit'
+      if category.update(category_params)
+        redirect_to category, notice: 'Category was successfully updated.'
+      else
+        render action: 'edit'
+      end
     end
   end
 
   def destroy
-    category.destroy
-    redirect_to categories_url, notice: 'Category was successfully destroyed.'
+    if current_user.admin?
+      category.destroy
+      redirect_to categories_url, notice: 'Category was successfully destroyed.'
+    else
+        redirect_to category, alert: 'You don\'t have permission to do this.'      
+    end
   end
 
   private
